Dissolve the remaining coverage grab-bag test files
tests/test_generator_coverage.py mixed tests for two unrelated modules
and tests/test_edge_case_coverage.py was a leftover catch-all. Their
tests now live with the other tests for the module they exercise:

- the normalize_openapi_31_* tests moved to
  tests/generators/test_cicerone_compat.py, which already tested that
  module
- the SchemasGenerator no-components/no-schemas tests moved to
  tests/generators/test_schemas_coverage.py (dropping the duplicate
  copy from test_edge_case_coverage.py)
- the APIGenerator regeneration and server-URL tests moved to
  tests/generators/framework/test_framework_generator_coverage.py,
  rewritten to build their specs with parse_spec_from_dict instead of
  tempfile round-trips
