+++ tests/generators/test_cicerone_compat.py
@@ -1,8 +1,9 @@
-"""Tests for cicerone_compat module to increase coverage."""
+"""Tests for the cicerone_compat module."""
 
 from unittest.mock import Mock
 
 from clientele.generators import cicerone_compat
+from clientele.generators.cicerone_compat import normalize_openapi_31_schema, normalize_openapi_31_spec
 
 
 def test_schema_to_dict_with_dict_input():
@@ -215,3 +216,161 @@ def test_get_pydantic_extra_with_key_present():
     result = cicerone_compat.get_pydantic_extra(obj, "test_key")
 
     assert result == "test_value"
+
+
+def test_normalize_openapi_31_schema_only_null_type():
+    """Test schema with only null type becomes nullable string."""
+    schema = {"type": ["null"]}
+    result = normalize_openapi_31_schema(schema)
+    assert result["type"] == "string"
+    assert result["nullable"] is True
+
+
+def test_normalize_openapi_31_schema_preserves_existing_nullable():
+    """Test that existing nullable: true is preserved."""
+    schema = {"type": "string", "nullable": True}
+    result = normalize_openapi_31_schema(schema)
+    assert result["nullable"] is True
+
+
+def test_normalize_openapi_31_schema_items():
+    """Test normalization of array items."""
+    schema = {"type": "array", "items": {"type": ["string", "null"]}}
+    result = normalize_openapi_31_schema(schema)
+    assert result["items"]["type"] == "string"
+    assert result["items"]["nullable"] is True
+
+
+def test_normalize_openapi_31_schema_allof():
+    """Test normalization of allOf schemas."""
+    schema = {"allOf": [{"type": ["string", "null"]}, {"minLength": 1}]}
+    result = normalize_openapi_31_schema(schema)
+    assert result["allOf"][0]["type"] == "string"
+    assert result["allOf"][0]["nullable"] is True
+
+
+def test_normalize_openapi_31_schema_oneof():
+    """Test normalization of oneOf schemas."""
+    schema = {"oneOf": [{"type": ["integer", "null"]}, {"type": "string"}]}
+    result = normalize_openapi_31_schema(schema)
+    assert result["oneOf"][0]["type"] == "integer"
+    assert result["oneOf"][0]["nullable"] is True
+
+
+def test_normalize_openapi_31_schema_anyof():
+    """Test normalization of anyOf schemas."""
+    schema = {"anyOf": [{"type": ["boolean", "null"]}, {"type": "number"}]}
+    result = normalize_openapi_31_schema(schema)
+    assert result["anyOf"][0]["type"] == "boolean"
+    assert result["anyOf"][0]["nullable"] is True
+
+
+def test_normalize_openapi_31_spec_components_schemas():
+    """Test normalization of schemas in components."""
+    spec = {
+        "openapi": "3.1.0",
+        "components": {"schemas": {"NullableString": {"type": ["string", "null"]}}},
+    }
+    result = normalize_openapi_31_spec(spec)
+    assert result["components"]["schemas"]["NullableString"]["type"] == "string"
+    assert result["components"]["schemas"]["NullableString"]["nullable"] is True
+
+
+def test_normalize_openapi_31_spec_request_body():
+    """Test normalization of request body schemas."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {
+            "/test": {
+                "post": {"requestBody": {"content": {"application/json": {"schema": {"type": ["string", "null"]}}}}}
+            }
+        },
+    }
+    result = normalize_openapi_31_spec(spec)
+    schema = result["paths"]["/test"]["post"]["requestBody"]["content"]["application/json"]["schema"]
+    assert schema["type"] == "string"
+    assert schema["nullable"] is True
+
+
+def test_normalize_openapi_31_spec_response_schemas():
+    """Test normalization of response schemas."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {
+            "/test": {
+                "get": {
+                    "responses": {"200": {"content": {"application/json": {"schema": {"type": ["integer", "null"]}}}}}
+                }
+            }
+        },
+    }
+    result = normalize_openapi_31_spec(spec)
+    schema = result["paths"]["/test"]["get"]["responses"]["200"]["content"]["application/json"]["schema"]
+    assert schema["type"] == "integer"
+    assert schema["nullable"] is True
+
+
+def test_normalize_openapi_31_spec_parameter_schemas():
+    """Test normalization of parameter schemas."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {
+            "/test": {
+                "get": {
+                    "parameters": [
+                        {
+                            "name": "test_param",
+                            "in": "query",
+                            "schema": {"type": ["boolean", "null"]},
+                        }
+                    ]
+                }
+            }
+        },
+    }
+    result = normalize_openapi_31_spec(spec)
+    param_schema = result["paths"]["/test"]["get"]["parameters"][0]["schema"]
+    assert param_schema["type"] == "boolean"
+    assert param_schema["nullable"] is True
+
+
+def test_normalize_openapi_31_spec_non_dict_path_item():
+    """Test that non-dict path items are skipped."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {
+            "/test": "not a dict",
+        },
+    }
+    result = normalize_openapi_31_spec(spec)
+    assert result["paths"]["/test"] == "not a dict"
+
+
+def test_normalize_openapi_31_spec_dollar_prefixed_keys():
+    """Test that $-prefixed keys (like $ref) are skipped."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {"/test": {"$ref": "#/components/paths/TestPath"}},
+    }
+    result = normalize_openapi_31_spec(spec)
+    assert result["paths"]["/test"]["$ref"] == "#/components/paths/TestPath"
+
+
+def test_normalize_openapi_31_spec_non_dict_response():
+    """Test that non-dict responses are skipped."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {"/test": {"get": {"responses": {"200": "not a dict"}}}},
+    }
+    result = normalize_openapi_31_spec(spec)
+    assert result["paths"]["/test"]["get"]["responses"]["200"] == "not a dict"
+
+
+def test_normalize_openapi_31_spec_non_dict_parameter():
+    """Test that non-dict parameters are handled."""
+    spec = {
+        "openapi": "3.1.0",
+        "paths": {"/test": {"get": {"parameters": ["not a dict"]}}},
+    }
+    result = normalize_openapi_31_spec(spec)
+    assert result["paths"]["/test"]["get"]["parameters"][0] == "not a dict"
